requirements.txt
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.13.4
blinker==1.7.0
Bottleneck==1.3.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
#cloud-init==25.1.4
colorama==0.4.6
#command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
#distro-info==1.7+build1
et-xmlfile==1.0.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
fonttools==4.46.0
fs==2.4.16
future==1.0.0
geocoder==1.38.1
googlemaps==4.10.0
greenlet==3.0.3
html5lib==1.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
#kiwisolver==0.0.0
kiwisolver==1.4.5
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
ldap3==2.9.1
django-auth-ldap==4.6.0
lxml==5.2.1
#lz4==4.0.2+dfsg
lz4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
mpmath
netifaces==0.11.0
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
packaging==24.0
pandas==2.1.4
pillow==10.2.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyasyncore==1.0.2
pycurl==7.45.3
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
pyinotify==0.9.6
PyJWT==2.7.0
PyMySQL==1.0.2
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
#python-apt==2.7.7
python-dateutil==2.8.2
python-redmine==2.5.0
pytz==2024.1
PyYAML==6.0.1
ratelim==0.1.6
redis==4.3.4
requests==2.31.0
rich==13.7.1
SciPy==1.11.4
service-identity==24.1.0
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
termcolor==2.4.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.4
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
Werkzeug==3.0.3
mysqlclient==2.2.0
Pillow==10.0.1
celery[redis]==5.4.0
redis==5.0.4
gunicorn==21.2.0
whitenoise==6.6.0
openpyxl==3.1.2
